Api and bll changes. Fixed DbContext and added new dal logic.

+++ TaskTracker.Dto/WorkTaskUserDto.cs
@@ -13,7 +13,7 @@ namespace TaskTracker.Dto
         public string City { get; set; }
         public string Mail { get; set; }
         public string PhoneNumber { get; set; }
-        public byte Role { get; set; }
+        public int Role { get; set; }
         public string Specialization { get; set; }
         public string AvatarPath { get; set; }
     }
